add SearchPanel to time-series-groups
progress on #1189
